Complete Authentication and CRUD APIs
Implemented the CRUD and Authentication APIs for tasks & projects.
