+++ Task-Management-Backend/.env
@@ -1,1 +1,2 @@
 JWT_SECRET="brother"
+PORT=5000
