+++ Task-Management-Backend/config/env-config.js
@@ -1,7 +1,6 @@
-const dotenv = require("dotenv");
+import dotenv from "dotenv";
 
 dotenv.config();
 
-module.exports = {
-    JWT_SECRET: process.env.JWT_SECRET,
-};
+export const JWT_SECRET = process.env.JWT_SECRET;
+export const PORT = process.env.PORT;
