--- Task-Management-Backend/middlewares/auth.middleware.js
@@ -1,1 +1,0 @@
-//Use you own middlewre and import to your own code as per need
