--- Task-Management-Backend/models/project.models.js
@@ -1,9 +1,0 @@
-import mongoose from "mongoose";
-
-const ProjectSchema = new mongoose.Schema({
-  name: { type: String, required: true },
-  description: { type: String, required: true }, 
-});
-
-const Project = mongoose.model("Project", ProjectSchema);
-export default Project;
