+++ Task-Management-Backend/models/task-models.js
@@ -1,12 +1,11 @@
-import mongoose from "mongoose"; 
+import { Schema, model } from "mongoose";
 
-const TaskSchema = new mongoose.Schema({
+const TaskSchema = new Schema({
   name: { type: String, required: true },
-  project: { type: mongoose.Schema.Types.ObjectId, ref: "Project" },
+  project: { type: Schema.Types.ObjectId, ref: "Project" },
   deadline: { type: Date },
   priority: { type: String, enum: ["Low", "Medium", "High"] },
   status: { type: String, enum: ["Pending", "Completed"], default: "Pending" },
 });
 
-const Task = mongoose.model("Task", TaskSchema);
-export default Task;
+export default model("Task", TaskSchema);
