+++ Task-Management-Backend/models/user-models.js
@@ -1,7 +1,7 @@
-const mongoose = require("mongoose");
-const bcrypt = require("bcrypt");
+import { Schema, model } from "mongoose";
+import { hash } from "bcrypt";
 
-const userSchema = new mongoose.Schema(
+const userSchema = new Schema(
     {
         email: { type: String, required: true, unique: true },
         username: { type: String, required: true },
@@ -12,8 +12,8 @@ const userSchema = new mongoose.Schema(
 
 userSchema.pre("save", async function () {
     if (this.isModified("password")) {
-        this.password = await bcrypt.hash(this.password, 12);
+        this.password = await hash(this.password, 12);
     }
 });
 
-module.exports = mongoose.model("User", userSchema);
+export default model("User", userSchema);
