+++ Task-Management-Backend/package.json
@@ -7,7 +7,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js",
-    "dev"   : "nodemon server.js"
+    "dev"   : "npx nodemon server.js"
   },
   "author": "Lakshmi Sowmya",
   "license": "ISC",
