+++ Task-Management-Backend/routes/auth-routes.js
@@ -1,0 +1,9 @@
+import { Router } from "express";
+import AuthController from "../controllers/auth-controllers.js";
+const router = Router();
+
+router.post("/register", AuthController.registerUser);
+
+router.post("/login", AuthController.loginUser);
+
+export default router;
