--- Task-Management-Backend/routes/auth.js
@@ -1,9 +1,0 @@
-const express = require("express");
-const AuthController = require("../controllers/auth-controller");
-const router = express.Router();
-
-router.post("/register", AuthController.registerUser);
-
-router.post("/login", AuthController.loginUser);
-
-module.exports = router;
