+++ Task-Management-Backend/server.js
@@ -1,8 +1,9 @@
 import express from "express";
-import projectsRouter from "./routes/project.routes.js";
-import tasksRouter from "./routes/tasks.routes.js";
-import healthcheckRouter from "./routes/healthcheck.routes.js";
-
+import projectsRouter from "./routes/project-routes.js";
+import tasksRouter from "./routes/task-routes.js";
+import healthcheckRouter from "./routes/healthcheck-routes.js";
+import authRouter from "./routes/auth-routes.js"
+import { PORT } from "./config/env-config.js";
 import cors from "cors";
 import connectDB from "./db/index.js";
 
@@ -12,17 +13,16 @@ app.use(express.json());
 app.use(cors());
 
 // Connect to MongoDB
-// connectDB();
-
-const PORT = 8000; // Change the port to 8000
+connectDB();
 
 app.listen(PORT, () => {
-  console.log(`Server is running on http://localhost:${PORT}`); // Log the HTTP link
+    console.log(`Server is running on http://localhost:${PORT}`); // Log the HTTP link
 });
 
 // Our project routes go here
-// http://localhost:8000/api/v1/healthcheck
+// http://localhost:5000/api/v1/healthcheck
 
 app.use("/api/v1/healthcheck", healthcheckRouter);
 app.use("/api/v1/projects", projectsRouter);
 app.use("/api/v1/tasks", tasksRouter);
+app.use("/api/v1/user", authRouter);
